+++ fn/__init__.py
@@ -2,4 +2,4 @@ from .stream import Stream
 from .underscore import shortcut as _
 from .func import F
 
-__version__ = "0.1.15"
+__version__ = "0.1.16"