+++ tests.py
@@ -264,6 +264,10 @@ class UnderscoreTestCase(unittest.TestCase):
         self.assertEqual("(x1, x2) => (x1 == x2)", str(_ == _))
         self.assertEqual("(x1, x2) => (x1 != x2)", str(_ != _))
 
+    def test_reverse_string_converting(self):
+        self.assertEqual("(x1, x2, x3) => ((x1 + x2) + x3)", str(_ + _ + _))
+        self.assertEqual("(x1, x2, x3) => (x1 + (x2 * x3))", str(_ + _ * _))
+
 class CompositionTestCase(unittest.TestCase):
 
     def test_composition(self):